+++ server/src/main/java/com/linbit/linstor/ErrorContextSupplier.java
@@ -8,4 +8,6 @@ public interface ErrorContextSupplier
      * Adds the returned String (if not null) to the ErrorReport.
      */
     @Nullable String getErrorContext();
+
+    boolean hasErrorContext();
 }